MFB-1676: one shape for Medicaid variants, plus the weatherization family
Medicaid now uses a subgroup per variant with a flat file per white label:
chip/, emergency/, disability/, presumptive/, limited/, for_children/ and the
other TX categories. That replaces three coexisting conventions (nested
per-white-label dirs, flat subgroup files, program-named files) and drops names
like co_medicaidadultwithdisability.py.

From the inheritance audit — a white-label calculator that subclasses another
calculator means an ungrouped family:

- cesn_leap subclasses CO's LEAP, so it joins cross_white_label/liheap/ as an
  eighth implementation.
- WAP is a federal DOE program with five implementations (CO, CESN, NC, TX, WA)
  and no shared base, so cross_white_label/weatherization/ now holds them.
- The MFB MedicareSavings base moves into msp/ beside its NC subclass.

Property-tax relief is deliberately NOT a family: cpcr, dptr, ks_k40h and
mo_pts are separate state programs with no federal program behind them, and
cesn_cpcr subclasses CO's because CESN reuses that program.

Also deletes cesn_calculators in co/energy_calculator/__init__.py — a registry
dict MFB-1675 missed because it sits two levels down. Nothing referenced it.
