+++ programs/framework/tests/test_presumptive_eligibility_resolution.py
@@ -21,10 +21,10 @@ from programs.framework.base import Eligibility
 from programs.programs.co.denver_sidewalk_rebate.calculator import DenverSidewalkRebate
 from programs.programs.co.denver_trash_rebate.calculator import DenverTrashRebate
 from programs.programs.co.energy_calculator.vehicle_exchange.calculator import EnergyCalculatorVehicleExchange
-from programs.programs.co.weatherization_assistance.calculator import WeatherizationAssistance
+from programs.programs.cross_white_label.weatherization.co import WeatherizationAssistance
 from programs.programs.cross_white_label.head_start.nc import NCHeadStart
 from programs.programs.nc.sun_bucks.calculator import SunBucks
-from programs.programs.tx.wap.calculator import TxWap
+from programs.programs.cross_white_label.weatherization.tx import TxWap
 from screener.models import Screen, WhiteLabel
 from screener.serializers import _write_current_benefits
 from screener.tests.helpers import seed_program
