+++ programs/framework/tests/test_receipt_contract.py
@@ -30,7 +30,7 @@ from programs.programs.cross_white_label.msp.base import Msp
 from programs.programs.cross_white_label.head_start.base import HeadStart
 from programs.programs.cross_white_label.early_head_start.base import EarlyHeadStart
 from programs.programs.cross_white_label.liheap.tx import TxCeap
-from programs.programs.cross_white_label.medicaid.disability.il_ilaabd import IlAabd
+from programs.programs.cross_white_label.medicaid.disability.il_aabd import IlAabd
 
 # The amount inputs predate the contract; everything else arrived with it.
 UNGATED_FIELDS = {"ssi", "tanf"}
