+++ programs/programs/co/energy_calculator/electric_affordability_xcel/calculator.py
@@ -1,8 +1,8 @@
 from programs.framework.base import Eligibility, ProgramCalculator
-from programs.programs.co.energy_calculator.energy_assistance.calculator import EnergyCalculatorEnergyAssistance
+from programs.programs.cross_white_label.liheap.cesn import EnergyCalculatorEnergyAssistance
 from programs.programs.co.energy_calculator.energy_outreach.calculator import EnergyCalculatorEnergyOutreach
 from programs.programs.co.energy_calculator.util import has_renter_expenses
-from programs.programs.co.energy_calculator.weatherization_assistance.calculator import (
+from programs.programs.cross_white_label.weatherization.cesn import (
     EnergyCalculatorWeatherizationAssistance,
 )
