+++ programs/programs/cross_white_label/medicaid/disability/tests/test_co_child.py
@@ -1,5 +1,5 @@
 """
-Income-test coverage for MedicaidAdultWithDisability (`awd_medicaid`).
+Income-test coverage for MedicaidChildWithDisability (`cwd_medicaid`).
 
 This is a non-MAGI Medicaid pathway — it counts unearned income SSI-style rather
 than through MAGI, so the Nurturing Futures direct cash payment is excluded from the
@@ -7,7 +7,7 @@ unearned side the same way it is for SSI itself. The verified income rules for t
 program cover "MAGI-based Health First Colorado and CHP+" only, which is why this
 pathway needed handling separately.
 
-Member income test: ((earned - $65) * 0.5) + (unearned - $20) <= 450% FPL.
+Household income test: (earned - $90 + unearned) * 0.67 <= 300% FPL.
 
 `calc_gross_income` is faked with a real exclude-honoring implementation so the tests
 pin eligibility outcomes rather than call signatures.
@@ -18,9 +18,9 @@ from unittest.mock import Mock
 from django.test import TestCase
 
 from programs.models import _FPL_DEFAULTS
-from programs.framework.base import MemberEligibility
-from programs.programs.cross_white_label.medicaid.disability.co_medicaidadultwithdisability import (
-    MedicaidAdultWithDisability,
+from programs.framework.base import Eligibility
+from programs.programs.cross_white_label.medicaid.disability.co_child import (
+    MedicaidChildWithDisability,
 )
 from programs.util import Dependencies
 
@@ -29,8 +29,9 @@ FPL_2025 = _FPL_DEFAULTS["2025"]
 EARNED_TYPES = ("wages", "selfEmployment")
 
 
-def make_income_fn(income):
+def make_calculator(income=None, household_size=1):
     """`income` is a dict of {income type: yearly amount}."""
+    income = income or {}
 
     def calc_gross_income(frequency, types, exclude=[]):
         total = 0
@@ -42,43 +43,30 @@ def make_income_fn(income):
                 total += amount
         return total
 
-    return calc_gross_income
-
-
-def make_calculator(household_size=1):
     program = Mock()
     program.year.as_dict.return_value = FPL_2025
 
     screen = Mock()
     screen.household_size = household_size
+    screen.calc_gross_income = Mock(side_effect=calc_gross_income)
 
-    return MedicaidAdultWithDisability(screen, program, {}, Dependencies())
-
-
-def make_member(income):
-    member = Mock()
-    member.age = 40
-    member.long_term_disability = True
-    member.visually_impaired = False
-    member.insurance.has_insurance_types.return_value = True
-    member.calc_gross_income = Mock(side_effect=make_income_fn(income))
-    return member
+    return MedicaidChildWithDisability(screen, program, {}, Dependencies())
 
 
-class TestAdultWithDisabilityIncome(TestCase):
+class TestChildWithDisabilityIncome(TestCase):
     def _run(self, income):
-        e = MemberEligibility(make_member(income))
-        make_calculator().member_eligible(e)
+        e = Eligibility()
+        make_calculator(income).household_eligible(e)
         return e.eligible
 
-    # 15,650 * 4.5 == 70,425 limit, against unearned less the $20 disregard
+    # 15,650 * 3 == 46,950 limit; countable income is 67% of the total
     def test_unearned_income_under_the_limit_is_eligible(self):
         self.assertTrue(self._run({"pension": 70_000}))
 
     def test_unearned_income_over_the_limit_is_ineligible(self):
         self.assertFalse(self._run({"pension": 71_000}))
 
-    def test_nurturing_futures_does_not_push_a_member_over_the_limit(self):
+    def test_nurturing_futures_does_not_push_a_household_over_the_limit(self):
         self.assertTrue(self._run({"pension": 69_000, "nurturingFutures": 7_200}))
 
     def test_a_counted_unearned_type_of_the_same_size_does_push_it_over(self):
