+++ programs/programs/il/pe/member.py
@@ -5,9 +5,9 @@ import programs.framework.pe_dependencies.household as household_dependency
 import programs.framework.pe_dependencies.member as member_dependency
 import programs.framework.pe_dependencies.spm as spm_dependency
 from programs.framework.pe_base import PolicyEngineMembersCalculator
-from programs.programs.cross_white_label.medicaid.il import IlMedicaid
 from programs.programs.cross_white_label.medicaid.base import Medicaid
 from programs.programs.cross_white_label.ssi.base import Ssi
+from programs.programs.cross_white_label.medicaid.il import IlMedicaid
 
 
 class IlBccp(PolicyEngineMembersCalculator):
@@ -64,43 +64,3 @@ class IlBccp(PolicyEngineMembersCalculator):
             return 400
 
         return 0
-
-
-class IlMpe(PolicyEngineMembersCalculator):
-    """
-    Illinois Medicaid Presumptive Eligibility (Pregnancy)
-
-    Eligibility criteria:
-        - Illinois resident
-        - Pregnant
-        - Meets income requirements for Medicaid Presumptive Eligibility
-        (as determined by PolicyEngine using the Medicaid income level -
-        approximately 200% of the FPL)
-        - Not already enrolled in Medicaid for the eligible individual
-    """
-
-    program_code = "il_mpe"
-
-    pe_name = "il_mpe_eligible"
-    pe_category = "people"
-
-    pe_inputs = [
-        member_dependency.AgeDependency,
-        *pe_dependency.irs_gross_income,
-        member_dependency.ExpectedChildrenPregnancyDependency,
-        household_dependency.IlStateCodeDependency,
-        member_dependency.PregnancyDependency,
-    ]
-
-    pe_outputs = [
-        member_dependency.IlMpeEligible,
-    ]
-
-    def member_value(self, member):
-        is_eligible = super().member_value(member)
-
-        has_medicaid = member.has_insurance_types(("medicaid",), strict=False)
-        if has_medicaid or not is_eligible:
-            return 0
-
-        return 1
