+++ programs/programs/tx/pe/tests/test_member.py
@@ -16,12 +16,12 @@ from programs.framework.pe_dependencies import household, member
 from programs.framework.pe_dependencies.household import TxStateCodeDependency
 from programs.programs.tx.pe.member import TxHarrisCountyRides, TxDart
 from programs.programs.cross_white_label.medicaid.base import Medicaid
-from programs.programs.cross_white_label.medicaid.tx.emergency_medicaid.calculator import TxEmergencyMedicaid
-from programs.programs.cross_white_label.medicaid.tx.for_children.calculator import TxMedicaidForChildren
-from programs.programs.cross_white_label.medicaid.tx.for_parents_and_caretakers.calculator import (
+from programs.programs.cross_white_label.medicaid.emergency.tx import TxEmergencyMedicaid
+from programs.programs.cross_white_label.medicaid.for_children.tx import TxMedicaidForChildren
+from programs.programs.cross_white_label.medicaid.for_parents_and_caretakers.tx import (
     TxMedicaidForParentsAndCaretakers,
 )
-from programs.programs.cross_white_label.medicaid.tx.for_pregnant_women.calculator import TxMedicaidForPregnantWomen
+from programs.programs.cross_white_label.medicaid.for_pregnant_women.tx import TxMedicaidForPregnantWomen
 from programs.programs.cross_white_label.wic.base import Wic
 from programs.programs.cross_white_label.wic.tx import TxWic
 from programs.programs.cross_white_label.ssi.base import Ssi
